### 2022-10-11 #4813 @osorensen
@annajenul, thanks for submitting. I'm the editor handling this submission now, and will start looking for reviewers. If you have any suggestions for reviewers, please let me know.

In the meantime, could you update your .bib file containing the references so that words that should be capitalized appear with capital letters in the paper? This is done by adding curly braces around the word, e.g. "{Bayes}" rather than "Bayes" in the second reference and "{R}" rather than "R" in the first reference.
