+++ vignettes/paper.bib
@@ -44,14 +44,14 @@
 
 
 @article{de2013mrmre,
-  title={mRMRe: an R package for parallelized mRMR ensemble feature selection},
-  author={De Jay, Nicolas and Papillon-Cavanagh, Simon and Olsen, Catharina and El-Hachem, Nehme and Bontempi, Gianluca and Haibe-Kains, Benjamin},
-  journal={Bioinformatics},
+  title={{mRMRe: an R package for parallelized mRMR ensemble feature selection}},
+  author={{De Jay, Nicolas and Papillon-Cavanagh, Simon and Olsen, Catharina and El-Hachem, Nehme and Bontempi, Gianluca and Haibe-Kains, Benjamin}},
+  journal={{Bioinformatics}},
   volume={29},
   number={18},
   pages={2365--2368},
   year={2013},
-  publisher={Oxford University Press}
+  publisher={{Oxford University Press}}
 }
 
 
