+++ vignettes/paper.bib
@@ -1,10 +1,3 @@
-@article{jenul2021user,
-  title={A User-Guided Bayesian Framework for Ensemble Feature Selection in Life Science Applications (UBayFS)},
-  author={Jenul, Anna and Schrunner, Stefan and Pilz, J{\"u}rgen and Tomic, Oliver},
-  journal={arXiv preprint arXiv:2104.14787},
-  year={2021}
-}
-
 @article{jenul2021rent,
   title={Rent—repeated elastic net technique for feature selection},
   author={Jenul, Anna and Schrunner, Stefan and Liland, Kristian Hovde and Indahl, Ulf Geir and Futsaether, Cecilia Marie and Tomic, Oliver},
@@ -16,30 +9,30 @@
 }
 
 @article{jenul2022ubayfs,
-  title={A user-guided Bayesian framework for ensemble feature selection in life science applications (UBayFS)},
-  author={Jenul, Anna and Schrunner, Stefan and Pilz, J{\"u}rgen and Tomic, Oliver},
-  journal={Machine Learning},
+  title={{A user-guided Bayesian framework for ensemble feature selection in life science applications (UBayFS)}},
+  author={{Jenul, Anna and Schrunner, Stefan and Pilz, J{\"u}rgen and Tomic, Oliver}},
+  journal={{Machine Learning}},
   pages={1--27},
   year={2022},
-  publisher={Springer}
+  publisher={{Springer}}
 }
 
 @article{scrucca2013ga,
-  title={GA: A package for genetic algorithms in R},
-  author={Scrucca, Luca},
-  journal={Journal of Statistical Software},
+  title={{GA: A package for genetic algorithms in R}},
+  author={{Scrucca, Luca}},
+  journal={{Journal of Statistical Software}},
   volume={53},
   pages={1--37},
   year={2013}
 }
 
 @article{you2022rdimtools,
-  title={Rdimtools: An r package for dimension reduction and intrinsic dimension estimation},
-  author={You, Kisung and Shung, Dennis},
-  journal={Software Impacts},
+  title={{Rdimtools: An R package for dimension reduction and intrinsic dimension estimation}},
+  author={{You, Kisung and Shung, Dennis}},
+  journal={{Software Impacts}},
   pages={100414},
   year={2022},
-  publisher={Elsevier}
+  publisher={{Elsevier}}
 }
 
 
@@ -56,17 +49,17 @@
 
 
   @Manual{gselection,
-    title = {GSelection: Genomic Selection},
-    author = {Sayanti Guha Majumdar and Anil Rai and Dwijesh Chandra Mishra},
+    title = {{GSelection: Genomic Selection}},
+    author = {{Sayanti Guha Majumdar and Anil Rai and Dwijesh Chandra Mishra}},
     year = {2019},
-    note = {R package version 0.1.0},
+    note = {{R package version 0.1.0}},
     url = {https://CRAN.R-project.org/package=GSelection},
   }
 
 @Manual{sgl,
-    title = {SGL: Fit a GLM (or Cox Model) with a Combination of Lasso and Group
-Lasso Regularization},
-    author = {Noah Simon and Jerome Friedman and Trevor Hastie and Rob Tibshirani},
+    title = {{SGL: Fit a GLM (or Cox Model) with a Combination of Lasso and Group
+Lasso Regularization}},
+    author = {{Noah Simon and Jerome Friedman and Trevor Hastie and Rob Tibshirani}},
     year = {2019},
     note = {R package version 1.3},
     url = {https://CRAN.R-project.org/package=SGL},
@@ -80,11 +73,11 @@ Lasso Regularization},
[…]
