+++ vignettes/paper.bib
@@ -1,5 +1,5 @@
 @article{jenul2021rent,
-  title={Rent—repeated elastic net technique for feature selection},
+  title={{RENT—Repeated Elastic Net Technique for Feature Selection}},
   author={Jenul, Anna and Schrunner, Stefan and Liland, Kristian Hovde and Indahl, Ulf Geir and Futsaether, Cecilia Marie and Tomic, Oliver},
   journal={IEEE Access},
   volume={9},
@@ -9,7 +9,7 @@
 }
 
 @article{jenul2022ubayfs,
-  title={{A user-guided Bayesian framework for ensemble feature selection in life science applications (UBayFS)}},
+  title={{A User-Guided Bayesian Framework for Ensemble Feature Selection in Life Science Applications (UBayFS)}},
   author={{Jenul, Anna and Schrunner, Stefan and Pilz, J{\"u}rgen and Tomic, Oliver}},
   journal={{Machine Learning}},
   pages={1--27},
@@ -18,7 +18,7 @@
 }
 
 @article{scrucca2013ga,
-  title={{GA: A package for genetic algorithms in R}},
+  title={{GA: A Package for Genetic Algorithms in R}},
   author={{Scrucca, Luca}},
   journal={{Journal of Statistical Software}},
   volume={53},
