### 2022-10-12 #4813 @osorensen
Thanks @dhvalden! There is not set time limit, but we hope you can start the review within 3 weeks. After that the authors might spend some time resolving issues, and then we hope you can come back and look at the revisions. Does that sound fine?
