### 2022-10-13 #4848 @editorialbot
[REVIEW]: UBayFS: An R Package for User Guided Feature Selection
**Submitting author:** @annajenul (<a href="http://orcid.org/0000-0002-6919-3483">Anna Jenul</a>)
**Repository:** https://github.com/annajenul/UBayFS
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
