+++ vignettes/paper.bib
@@ -5,7 +5,8 @@
   volume={9},
   pages={152333--152346},
   year={2021},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/ACCESS.2021.3126429}
 }
 
 @article{jenul2022ubayfs,
@@ -14,7 +15,9 @@
   journal={{Machine Learning}},
   pages={1--27},
   year={2022},
-  publisher={{Springer}}
+  publisher={{Springer}},
+  doi={10.1007/s10994-022-06221-9}
+
 }
 
 @article{scrucca2013ga,
@@ -23,7 +26,8 @@
   journal={{Journal of Statistical Software}},
   volume={53},
   pages={1--37},
-  year={2013}
+  year={2013},
+  doi={10.18637/jss.v053.i04}
 }
 
 @article{you2022rdimtools,
@@ -32,7 +36,8 @@
   journal={{Software Impacts}},
   pages={100414},
   year={2022},
-  publisher={{Elsevier}}
+  publisher={{Elsevier}},
+  doi={10.1016/j.simpa.2022.100414}
 }
 
 
@@ -44,7 +49,8 @@
   number={18},
   pages={2365--2368},
   year={2013},
-  publisher={{Oxford University Press}}
+  publisher={{Oxford University Press}},
+  doi={10.1093/bioinformatics/btt383}
 }
 
 
@@ -73,7 +79,8 @@ Lasso Regularization}},
     pages = "124--139",
     year = "2017",
     issn = "0950-7051",
-    author = {{B. Seijo-Pardo and I. Porto-Díaz and V. Bolón-Canedo and A. Alonso-Betanzos}}
+    author = {{B. Seijo-Pardo and I. Porto-Díaz and V. Bolón-Canedo and A. Alonso-Betanzos}},
+    doi={10.1016/j.knosys.2016.11.017}
 }
 
 @inproceedings{ida:fSparseGlasso,
