### 2022-11-01 annajenul/UBayFS#7 @aaronpeikert
Consider renaming default branch to  main
Just a minor thing really but I think it makes sense:
https://github.com/github/renaming

Related:
https://github.com/openjournals/joss-reviews/issues/4848
