### 2022-11-01 annajenul/UBayFS#8 @aaronpeikert
NAMESPACE is missing
Installing with:

`remotes::install_github("annajenul/UBayFS", build_vignettes = TRUE)`

Fails:

```
The downloaded source packages are in
	‘/tmp/RtmpifgaPc/downloaded_packages’
✔  checking for file ‘/tmp/RtmpifgaPc/remotes408bd37207935/annajenul-UBayFS-f48ca08/DESCRIPTION’ ...
─  preparing ‘UBayFS’:
         -----------------------------------
─  installing *source* package ‘UBayFS’ ...
   ** using staged installation
   ERROR: a 'NAMESPACE' file is required
─  removing ‘/tmp/RtmpF4XUwT/Rinst4f4b820f9a425/UBayFS’
         -----------------------------------
   ERROR: package installation failed
Error: Failed to install 'UBayFS' from GitHub:
  ! System command 'R' failed
```
