### 2022-11-01 annajenul/UBayFS#9 @aaronpeikert
installation requires suggested packages
The suggested installation method also builds the vignettes, which requires `prettydocs` which is not a required package:

`remotes::install_github("annajenul/UBayFS", build_vignettes = TRUE)`
