### 2022-11-01 annajenul/UBayFS#11 @aaronpeikert
Test failure in build models
https://github.com/annajenul/UBayFS/blob/f48ca0873ee9d617db9417bd030bcef6fcdbbc77/tests/testthat/test_build.R#L72-L105
Results in test failure:

```
`actual[1:9]`:  9 4  9 10 1 1 10 10 0
`expected[1:9]`: 10 0 10 10 0 0 10 10 0

  `actual[11:17]`: 0 0 0  2 0 0 0
`expected[11:17]`: 0 0 0 10 0 0 0

  `actual[19:30]`: 0 0 10 7 10 10 1 0 0  6 and 2 more...
`expected[19:30]`: 0 0 10 0 10 10 0 0 0 10           ...
```
