### 2022-11-01 annajenul/UBayFS#13 @aaronpeikert
XQuartz is an unstated dependency / is not installed automatically
Tests on MacOS indicate that an unstated dependency is xquartz:

https://github.com/aaronpeikert/UBayFS/actions/runs/3368031028/jobs/5586118641
