### 2022-11-01 annajenul/UBayFS#14 @aaronpeikert
setup CI
I think having a CI really helps a lot and its pretty easy:

`usethis::use_github_action_check_standard()`

Revealed for example #13
