### 2022-11-01 #4848 @aaronpeikert
I have only gotten around to take a quick look at the installation, dependencies, and automated testing:

- [ ] https://github.com/annajenul/UBayFS/issues/8
- [ ] https://github.com/annajenul/UBayFS/issues/9
- [ ] https://github.com/annajenul/UBayFS/issues/11
- [ ] https://github.com/annajenul/UBayFS/issues/13
- [ ] https://github.com/annajenul/UBayFS/issues/14

I'll check back in two weeks to do the rest.
