### 2022-11-02 #4848 @osorensen
Thanks @aaronpeikert!

FYI @annajenul, you're welcome to start addressing these issues right away, as the review process in JOSS is supposed to be iterative.
