+++ .github/workflows/R-CMD-check.yaml
@@ -47,3 +47,6 @@ jobs:
       - uses: r-lib/actions/check-r-package@v2
         with:
           upload-snapshots: true
+
+      - if:runner.os == "macOS"
+        run: brew install --cask xquartz
