+++ .github/workflows/R-CMD-check.yaml
@@ -59,7 +59,7 @@ jobs:
         env:
           RGL_USE_NULL: TRUE
           DISPLAY: ":99"
-        run: Rscript -e "rcmdcheck::rcmdcheck(args = '--no-manual', error_on = 'warning', check_dir = 'check')"
+        run: Rscript -e "rcmdcheck::rcmdcheck(args = '--no-manual', error_on = 'error', check_dir = 'check')"
 
 
       - uses: r-lib/actions/check-r-package@v2
