+++ .github/workflows/R-CMD-check.yaml
@@ -26,6 +26,7 @@ jobs:
 
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
+      R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
       R_KEEP_PKG_SOURCE: yes
 
     steps:
@@ -50,10 +51,10 @@ jobs:
           brew install xquartz --cask
 
 
-      - uses: r-lib/actions/setup-r-dependencies@v2
-        with:
-          extra-packages: any::rcmdcheck
-          needs: check
+     # - uses: r-lib/actions/setup-r-dependencies@v2
+    #    with:
+    #      extra-packages: any::rcmdcheck
+    #      needs: check
 
       - name: Check
         env:
