### 2022-11-04 9d9ce83
install cmd check
