+++ .github/workflows/R-CMD-check.yaml
@@ -25,50 +25,61 @@ jobs:
           - {os: ubuntu-latest,   r: 'oldrel-1'}
 
     env:
-      GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
-      R_KEEP_PKG_SOURCE: yes
+      CRAN: ${{ matrix.config.cran }}
+
 
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v1
 
-      - uses: r-lib/actions/setup-pandoc@v2
-
-      - uses: r-lib/actions/setup-r@v2
+      - uses: r-lib/actions/setup-r@master
         with:
           r-version: ${{ matrix.config.r }}
-          http-user-agent: ${{ matrix.config.http-user-agent }}
-          use-public-rspm: true
 
+      - uses: r-lib/actions/setup-pandoc@master
 
       - name: Query dependencies
         run: Rscript -e "install.packages('remotes')" -e "saveRDS(remotes::dev_package_deps(dependencies = TRUE), 'depends.Rds', version = 2)"
 
-      - name: Install qpdf / Linux
-        if: runner.os == 'Linux'
-        run: |
-          sudo apt-get install qpdf
+      - name: "[Linux & macOS] Cache R packages"
+        if: runner.os != 'Windows'
+        uses: actions/cache@v1
+        with:
+          path: ${{ env.R_LIBS_USER }}
+          key: ${{ runner.os }}-r-${{ matrix.config.r }}-${{ hashFiles('depends.Rds') }}
+          restore-keys: ${{ runner.os }}-r-${{ matrix.config.r }}-
 
-      - name: Install qpdf / Mac OS
-        if: runner.os == 'macOS'
+      - name: "[Linux] Install system dependencies"
+        if: runner.os == 'Linux'
+        env:
+          RHUB_PLATFORM: linux-x86_64-ubuntu-gcc
         run: |
-          brew install xquartz --cask
+          Rscript -e "remotes::install_github('r-hub/sysreqs')"
+          sysreqs=$(Rscript -e "cat(sysreqs::sysreq_commands('DESCRIPTION'))")
+          sudo -s eval "$sysreqs"
+
+      - name: "[macOS] Install system dependencies"
+        if: runner.os == 'macOS'
+        run:
+          brew cask install xquartz
 
       - name: Install dependencies
         run: Rscript -e "library(remotes)" -e "update(readRDS('depends.Rds'))" -e "remotes::install_cran('rcmdcheck')"
-
-     # - uses: r-lib/actions/setup-r-dependencies@v2
-    #    with:
-    #      extra-packages: any::rcmdcheck
-    #      needs: check
 
       - name: Check
         env:
           RGL_USE_NULL: TRUE
           DISPLAY: ":99"
-        run: Rscript -e "rcmdcheck::rcmdcheck(args = '--no-manual', error_on = 'error', check_dir = 'check')"
+        run: Rscript -e "rcmdcheck::rcmdcheck(args = '--no-manual', error_on = 'warning', check_dir = 'check')"
 
-
-      - uses: r-lib/actions/check-r-package@v2
+      - name: Upload check results
+        if: failure()
+        uses: actions/upload-artifact@master
         with:
-          upload-snapshots: true
+          name: ${{ runner.os }}-r${{ matrix.config.r }}-results
+          path: check
+
+      - name: Test coverage
+        if: matrix.config.os == 'ubuntu-16.04' && matrix.config.r == '3.6'
+        run: |
+          Rscript  -e "install.packages('covr')" -e 'covr::codecov(token = "${{secrets.CODECOV_TOKEN}}")'
