+++ .github/workflows/R-CMD-check.yaml
@@ -74,7 +74,7 @@ jobs:
         env:
           RGL_USE_NULL: TRUE
           DISPLAY: ":99"
-        run: Rscript -e "rcmdcheck::rcmdcheck(args = '--no-manual', error_on = 'warning', check_dir = 'check')"
+        run: Rscript -e "rcmdcheck::rcmdcheck(args = '--no-manual', error_on = 'error', check_dir = 'check')"
 
       - name: Upload check results
         if: failure()
