+++ .github/workflows/R-CMD-check.yaml
@@ -65,7 +65,7 @@ jobs:
       - name: "[macOS] Install system dependencies"
         if: runner.os == 'macOS'
         run:
-          brew cask install xquartz
+          brew install xquartz --cask
 
       - name: Install dependencies
         run: Rscript -e "library(remotes)" -e "update(readRDS('depends.Rds'))" -e "remotes::install_cran('rcmdcheck')"
