### 2022-11-05 annajenul/UBayFS#14 @annajenul
Thanks for the suggestion! I added a CI
