+++ tests/testthat/test_build.R
@@ -80,7 +80,7 @@ test_that("ensemble is trained correctly",{
     constraints = c
   )
   expect_equal(unname(model$ensemble.params$output$counts),
-               c(0, 10, 74, 0, 0, 0, 100, 100, 0, 0, 0, 0, 6, 100, 0, 0, 6, 0, 0, 0, 100, 86, 100, 100, 2, 16, 100, 100, 0, 0))
+               c(0,11,81,0,0,0,100,100,0,0,0,0,1,100,0,0,3,0,0,0,100,87,100,100,3,14,100,100,0,0))
 
   set.seed(1)
   model <- build.UBaymodel(
@@ -102,6 +102,6 @@ test_that("ensemble is trained correctly",{
     method = "dtree",
   )
   expect_equal(unname(model$ensemble.params$output$counts),
-               c(10, 0, 10, 10, 1, 1, 9, 8, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 10, 2, 10, 10, 1, 1, 6, 10, 0, 1))
+               c(10, 0, 10, 10, 0, 1, 6, 6, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 10, 0, 10, 10, 0, 6, 4, 9, 2, 1))
 
 })
