### 2022-11-09 annajenul/UBayFS#13 @annajenul
Thank you very much for reporting this issue. I added a note for macOS users to the installation section of the README file.
