### 2022-11-09 annajenul/UBayFS#9 @annajenul
Thank you, we moved prettydocs to the required packages.
