### 2022-11-10 e207afd
update UBayFS build function
