+++ R/buildUBaymodel.R
@@ -6,7 +6,7 @@
 #' @param M the number of elementary models to be trained in the ensemble
 #' @param tt_split the ratio of samples drawn for building an elementary model (train-test-split)
 #' @param nr_features number of features to select in each elementary model
-#' @param method a vector denoting the method(s) used as elementary models; options: `mRMR`, `laplace` (Laplacian score)
+#' @param method a vector denoting the method(s) used as elementary models; options: `mRMR`, `laplace` (Laplacian score) Also self-defined functions are possible methods; they must have the arguments X (data), y (target), n (number of features) and name (name of the function). For more details see examples.
 #' @param prior_model a string denoting the prior model to use; options: `dirichlet`, `wong`, `hankin`; `hankin` is the most general prior model, but also the most time consuming
 #' @param weights the vector of user-defined prior weights for each feature
 #' @param lambda a positive scalar denoting the overall strength of the constraints
@@ -16,6 +16,7 @@
 #' @param popsize size of the initial population of the genetic algorithm for model optimization
 #' @param maxiter maximum number of iterations of the genetic algorithm for model optimization
 #' @param shiny TRUE indicates that the function is called from Shiny dashboard
+#' @param ... additional arguments
 #' @return a `UBaymodel` object containing the following list elements:
 #' \itemize{
 #'   \item `data` - the input dataset
@@ -42,6 +43,24 @@
 #'                      weights = w
 #' )
 #'
+#' # use a function computing a decision tree as input
+#' library("rpart")
+#' decision_tree <- function(X, y, n, name = "tree"){
+#' rf_data = as.data.frame(cbind(y, X))
+#' colnames(rf_data) <- make.names(colnames(rf_data))
+#' tree = rpart::rpart(y~., data = rf_data)
+#' return(list(ranks= which(colnames(X) %in% names(tree$variable.importance)[1:n]),
+#'            name = name))
+#' }
+#'
+#' model <- build.UBaymodel(
+#'                      data = bcw$data,
+#'                      target = bcw$labels,
+#'                      constraints = c,
+#'                      weights = w,
+#'                      method = decision_tree
+#' )
+#'
 #' # include block-constraints
 #' c_block <- buildConstraints(constraint_types = "max_size",
 #'                             constraint_vars = list(2),
@@ -56,11 +75,8 @@
 #'                      weights = w
 #' )
 #' @import Rdimtools
-#' @import glmnet
 #' @import mRMRe
 #' @import shiny
-#' @import caret
-#' @importFrom GSelection feature.selection
 #' @export
 
 build.UBaymodel = function(data,
@@ -77,12 +93,15 @@ build.UBaymodel = function(data,
                            optim_method = "GA",
                            popsize = 50,
                            maxiter = 100,
-                           shiny = FALSE){
+                           shiny = FALSE,
+                           ...){
 
   # check input
   if(!is.matrix(data)){
     data = as.matrix(data)
   }
+  if(is.function(method)){method = list(method)}
+
   if(nrow(data) != length(target)){
     stop("Error: number of labels must match number of data rows")
   }
@@ -95,10 +114,11 @@ build.UBaymodel = function(data,
   else if(tt_split < 0.5 | tt_split > 0.99){
     warning("Warning: tt_split should not be outside [0.5,0.99]")
   }
-  if(!all(method %in% c("mRMR", "mrmr", "Laplacian score", "laplace", "lasso", "LASSO", "fisher", "Fisher", "RFE", "rfe", "hsic", "HSIC",
[…]
