### 2022-11-11 7f47afd
update vignette, shiny dashboard; handling dependencies for shiny dashboard
