+++ NAMESPACE
@@ -25,10 +25,13 @@ import(gridExtra)
 import(hyper2)
 import(mRMRe)
 import(shiny)
+import(tcltk)
 importFrom(DirichletReg,ddirichlet)
 importFrom(GA,ga)
 importFrom(matrixStats,logSumExp)
 importFrom(methods,is)
 importFrom(stats,cor)
 importFrom(utils,combn)
+importFrom(utils,install.packages)
+importFrom(utils,installed.packages)
 importFrom(utils,str)
