### 2022-11-11 cfe866c
modification in runInteractive
