### 2022-11-11 2906a75
replace tcltk with select.list from utils in the run_interactive function
