+++ NAMESPACE
@@ -25,7 +25,6 @@ import(gridExtra)
 import(hyper2)
 import(mRMRe)
 import(shiny)
-import(tcltk)
 importFrom(DirichletReg,ddirichlet)
 importFrom(GA,ga)
 importFrom(matrixStats,logSumExp)
@@ -34,4 +33,5 @@ importFrom(stats,cor)
 importFrom(utils,combn)
 importFrom(utils,install.packages)
 importFrom(utils,installed.packages)
+importFrom(utils,select.list)
 importFrom(utils,str)
