+++ man/print.UBaymodel.Rd
@@ -1,0 +1,38 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/printUBaymodel.R
+\name{print.UBaymodel}
+\alias{print.UBaymodel}
+\alias{printResults}
+\alias{summary.UBaymodel}
+\alias{plot.UBaymodel}
+\title{Print a UBayFS model}
+\usage{
+\method{print}{UBaymodel}(x, ...)
+
+printResults(model)
+
+\method{summary}{UBaymodel}(object, ...)
+
+\method{plot}{UBaymodel}(x, ...)
+}
+\arguments{
+\item{x}{a `UBaymodel` object created using \link{build.UBaymodel}}
+
+\item{...}{additional print parameters}
+
+\item{model}{a `UBaymodel` object created using \link{build.UBaymodel} after training}
+
+\item{object}{a `UBaymodel` object created using \link{build.UBaymodel}}
+}
+\description{
+Print details of a UBayFS model.
+}
+\section{Functions}{
+\itemize{
+\item \code{printResults()}: Display and summarize the results of UBayFS after feature selection.
+
+\item \code{summary(UBaymodel)}: A summary of a UBayFS model
+
+\item \code{plot(UBaymodel)}: A barplot of an UBayFS model containing prior weights, ensemble counts and the selected features.
+
+}}
