+++ man/checkConstraints.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/buildConstraints.R
+\name{checkConstraints}
+\alias{checkConstraints}
+\title{Checks whether a list object implements proper UBayFS user constraints}
+\usage{
+checkConstraints(x)
+}
+\arguments{
+\item{x}{a `UBaymodel` object}
+}
+\value{
+boolean value
+}
+\description{
+Checks whether a list object implements proper UBayFS user constraints
+}
