+++ man/runInteractive.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/runInteractive.R
+\name{runInteractive}
+\alias{runInteractive}
+\title{Run an interactive shiny app for demonstration}
+\usage{
+runInteractive()
+}
+\description{
+Start a shiny application in the browser.
+}
