### 2022-11-15 #4848 @aaronpeikert
@annajenul I'd like to review the software's functionality more in-depth (by, e.g. simulating some data), but this requires more time than I have this week. Do you prefer that I take an in-depth look, which might lengthen the review by 2-3 weeks or a more cursory look (fitting some models to data where I do not know the true answers)?
