+++ DESCRIPTION
@@ -40,3 +40,4 @@ Suggests:
     glmnet
 Config/testthat/edition: 3
 VignetteBuilder: knitr
+URL: https://annajenul.github.io/UBayFS/
