+++ pkgdown.yml
@@ -1,0 +1,11 @@
+pandoc: 2.19.2
+pkgdown: 2.0.6
+pkgdown_sha: ~
+articles:
+  BFS_UBayFS: BFS_UBayFS.html
+  UBayFS: UBayFS.html
+last_built: 2022-11-15T14:16Z
+urls:
+  reference: https://annajenul.github.io/UBayFS/reference
+  article: https://annajenul.github.io/UBayFS/articles
+
