+++ _pkgdown.yml
@@ -1,4 +1,8 @@
 url: https://annajenul.github.io/UBayFS/
 template:
   bootstrap: 5
+navbar:
+  structure:
+    left:  [intro, reference, articles, tutorials, news]
+    right: [search, github]
 
