+++ _pkgdown.yml
@@ -1,5 +1,3 @@
 url: https://annajenul.github.io/UBayFS/
 template:
-  bootstrap: 5
-
-
+  bootstrap: 6
