+++ _pkgdown.yml
@@ -1,3 +1,3 @@
 url: https://annajenul.github.io/UBayFS/
 template:
-  bootstrap: 6
+  bootstrap: 5
