+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   BFS_UBayFS: BFS_UBayFS.html
   UBayFS: UBayFS.html
-last_built: 2022-11-15T17:05Z
+last_built: 2022-11-15T17:15Z
 urls:
   reference: https://annajenul.github.io/UBayFS/reference
   article: https://annajenul.github.io/UBayFS/articles
