+++ _pkgdown.yml
@@ -1,3 +1,8 @@
 url: https://annajenul.github.io/UBayFS/
-template:
+package: tidytemplate
   bootstrap: 5
+  
+  bslib:
+    primary: "#096B72"
+    navbar-background: "#e6f3fc"
+  trailing_slash_redirect: true
