+++ .github/workflows/pkgdown.yaml
@@ -28,24 +28,13 @@ jobs:
         with:
           use-public-rspm: true
 
-      - name: Install system dependencies
-        if: runner.os == 'Linux'
-        run: |
-          sudo apt update
-          sudo apt-get update
-          sudo apt-get -y install libssh-dev #package ssh dependency
-          while read -r cmd
-          do
-            eval sudo $cmd
-          done < <(Rscript -e 'writeLines(remotes::system_requirements("ubuntu", "20.04"))')
-
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
           extra-packages: any::pkgdown, local::.
           needs: website
 
       - name: Build site
-        run: pkgdown::build_site_github_pages(new_process = TRUE, install = FALSE)
+        run: pkgdown::build_site_github_pages(new_process = FALSE, install = FALSE)
         shell: Rscript {0}
 
       - name: Deploy to GitHub pages 🚀
