+++ README.md
@@ -3,11 +3,9 @@
   [![R-CMD-check](https://github.com/annajenul/UBayFS/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/annajenul/UBayFS/actions/workflows/R-CMD-check.yaml)
 <!-- badges: end -->
 
-<<<<<<< HEAD
-# UBayFS <img src="vignettes/logo.png" style="float:right; height:200px;" />
+
+# UBayFS <img src="vignettes/logo.png" align="right" height:200 />
 =======
-# UBayFS <img src="inst/shinyApp/UBayFSInteractive/www/logo.png" align="right" height=200 />
->>>>>>> 9987877ab6cae46fa0a38a916371200f25b598ba
 
 
 The UBayFS package implements the framework proposed in the article [Jenul et al. (2022)](https://link.springer.com/article/10.1007/s10994-022-06221-9), together with an interactive Shiny dashbord, which makes UBayFS applicable to R-users with different levels of expertise. UBayFS is an ensemble feature selection technique embedded in a Bayesian statistical framework. The method combines data and user knowledge, where the first is extracted via data-driven ensemble feature selection. The user can control the feature selection by assigning prior weights to features and penalizing specific feature combinations. In particular, the user can define a maximal number of selected features and must-link constraints (features must be selected together) or cannot-link constraints (features must not be selected together). Using relaxed constraints, a parameter $\rho$ regulates the penalty shape. Hence, violation of constraints can be valid but leads to a lower target value of the feature set that is derived from the violated constraints. UBayFS can be used for common feature selection and also for block feature selection.
