+++ vignettes/paper.bib
@@ -54,21 +54,31 @@
 }
 
 
-  @Manual{gselection,
-    title = {{GSelection: Genomic Selection}},
-    author = {{Sayanti Guha Majumdar and Anil Rai and Dwijesh Chandra Mishra}},
-    year = {2019},
-    note = {{R package version 0.1.0}},
-    url = {https://CRAN.R-project.org/package=GSelection},
-  }
+@Manual{gselection,
+  title = {{GSelection: Genomic Selection}},
+  author = {{Sayanti Guha Majumdar and Anil Rai and Dwijesh Chandra Mishra}},
+  year = {2019},
+  note = {{R package version 0.1.0}},
+  url = {https://CRAN.R-project.org/package=GSelection},
+}
 
-@Manual{sgl,
-    title = {{SGL: Fit a GLM (or Cox Model) with a Combination of Lasso and Group
-Lasso Regularization}},
-    author = {{Noah Simon and Jerome Friedman and Trevor Hastie and Rob Tibshirani}},
-    year = {2019},
-    note = {R package version 1.3},
-    url = {https://CRAN.R-project.org/package=SGL},
+@Article{boruta,
+  title = {Feature Selection with the {Boruta} Package},
+  author = {Miron B. Kursa and Witold R. Rudnicki},
+  journal = {Journal of Statistical Software},
+  year = {2010},
+  volume = {36},
+  number = {11},
+  pages = {1--13},
+  url = {https://doi.org/10.18637/jss.v036.i11},
+}
+
+@Manual{caret,
+    title = {caret: Classification and Regression Training},
+    author = {Max Kuhn},
+    year = {2022},
+    note = {R package version 6.0-93},
+    url = {https://CRAN.R-project.org/package=caret},
   }
 
 
@@ -83,11 +93,3 @@ Lasso Regularization}},
     doi={10.1016/j.knosys.2016.11.017}
 }
 
-@inproceedings{ida:fSparseGlasso,
- author = {{Ida, Yasutoshi and Fujiwara, Yasuhiro and Kashima, Hisashi}},
- booktitle = {{Advances in Neural Information Processing Systems}},
- publisher = {{Curran Associates, Inc.}},
- title = {{Fast Sparse Group Lasso}},
- volume = {32},
- year = {2019}
-}
