### 2022-11-26 6d1473c
updated vignette layout
