+++ vignettes/UBayFS.Rmd
@@ -15,12 +15,20 @@ vignette: >
 ---
 
 <style>
-
   body {
   text-align: justify;
   padding: 1em}
+  header:after {
+    content: "";
+    position: absolute;
+    top: 55px;
+    right: 40px;
+    width: 120px;
+    height: 120px;
+    background-image:url("logo.png");
+    background-size:120px;
+  }
 </style>
-
 
 ```{r, include = FALSE}
 library(knitr)
@@ -34,7 +42,7 @@ knitr::opts_chunk$set(fig.align = 'center')
 ```
 
 
-## Introduction
+# Introduction
 The UBayFS package implements the framework proposed in [@jenul2022ubayfs], together with an interactive Shiny dashboard, which makes UBayFS applicable to R-users with different levels of expertise. UBayFS is an ensemble feature selection technique embedded in a Bayesian statistical framework. The method combines data and user knowledge, where the first is extracted via data-driven ensemble feature selection. The user can control the feature selection by assigning prior weights to features and penalizing specific feature combinations. In particular, the user can define a maximum number of selected features and must-link constraints (features must be selected together) or cannot-link constraints (features must not be selected together). A parameter $\rho$ regulates the shape of a penalty term accounting for side constraints, where feature sets that violate constraints lead to a lower target value. 
 
 In this vignette, we use the Breast Cancer Wisconsin dataset [@wolberg:wisconsin] for demonstration. Specifically, the dataset consists of 569 samples and 30 features and can be downloaded as a demo dataset by calling ``data(bcw)``. The dataset describes a classification problem, where the aim is to distinguish between malignant and benign cancer based on image data. Features are derived from 10 image characteristics, where each characteristic is represented by three features (summary statistics) in the dataset. For instance, the characteristic *radius* is represented by features *radius mean*, *radius standard deviation*, and *radius worst*.
@@ -46,7 +54,7 @@ UBayFS is implemented via a core S3-class ``UBaymodel``, along with help functio
 include_graphics("UBay_UML.jpg")
 ``` 
 
-### Requirements and dependencies
+## Requirements and dependencies
 - R (>= 3.5.0)
 - GA
 - matrixStats
@@ -83,19 +91,19 @@ library(UBayFS)
 data(bcw)
 ```
 
-## Background
+# Background
 This section summarizes the core parts of UBayFS, where a central part is Bayes' Theorem for two random variables $\boldsymbol{\theta}$ and $\boldsymbol{y}$:
 $$p(\boldsymbol{\theta}|\boldsymbol{y})\propto p(\boldsymbol{y}|\boldsymbol{\theta})\cdot p(\boldsymbol{\theta}),$$ where $\boldsymbol{\theta}$ represents an importance parameter of single features and $\boldsymbol{y}$ collects evidence about $\boldsymbol{\theta}$ from an ensemble of elementary feature selectors. In the following, the concept will be outlined. 
 
-### Ensemble feature selection as likelihood
+## Ensemble feature selection as likelihood
[…]
