### 2022-11-26 3fed40a
updated logo on vignette
