### 2022-11-26 c493962
updated vignette layout
