### 2022-12-06 annajenul/UBayFS#22 @dhvalden
RMarkdown required but not listed in dependencies
Under Fedora Linux 37, R version 4.2.2 (2022-10-31), suggested package installation `remotes::install_github("annajenul/UBayFS", build_manual = TRUE, build_vignettes = TRUE)`.

Error message:
```
Running `R CMD build`...
* checking for file ‘/tmp/Rtmp5V5S0y/remotes118345827dd25/annajenul-UBayFS-c493962/DESCRIPTION’ ... OK
* preparing ‘UBayFS’:
* creating vignettes ... ERROR
--- re-building ‘BFS_UBayFS.Rmd’ using rmarkdown
Error: processing vignette 'BFS_UBayFS.Rmd' failed with diagnostics:
Pandoc is required to build R Markdown vignettes but not available. Please make sure it is installed.
--- failed re-building ‘BFS_UBayFS.Rmd’
```
