### 2022-12-06 #4848 @dhvalden
So far the installation seems to be running ok with the exception of this issue:

- [ ] https://github.com/annajenul/UBayFS/issues/22#issue-1479897544
