### 2022-12-07 49b5245
added Pandoc installation guideline
