+++ README.md
@@ -76,6 +76,13 @@ The development version of the package can be installed with:
 
 If you use a macOS operator system, make sure you have [XQuartz](https://www.xquartz.org/) installed.
 
+To build the vignettes, Pandoc is required. It may happen that Pandoc is missing on your computer, or that the version is too old. Then the installation will return the error
+    
+    ```
+    Pandoc is required to build R Markdown vignettes but not available. Please make sure it is installed.
+    ```
+An installation guide for Pandoc on different operation systems is provided [here](https://pandoc.org/installing.html).
+
 Contributing
 ------------
 Your contribution to UBayFS is very welcome! 
