### 2022-12-07 annajenul/UBayFS#22 @dhvalden
Just a note, I can easily bypass the issue by just not building the vignettes, so my suggested solution would just to change the suggested installation to not build vignettes by default.
