### 2022-12-07 annajenul/UBayFS#23 @dhvalden
UBayFS::train very verbose
The solution is to pass `monitor=FALSE` to the `ga()` fuction in `train_GA()`
Additionally, the message `"Running Genetic Algorithm"` can be printed to `stderr` using `message()`, which will get rid of the `[1]` character at the beginning of the message.
A similar behaviour can be achieved with `cat()`, but sending the output to `stdout`. Mind you that `cat()` requires to add newline character at the end `\n`
