### 2022-12-07 #4848 @dhvalden
As an update, I can confirm the core functionality with the exception of the shiny dashboard. I listed a couple of very minor improvements in here:
- [ ] https://github.com/annajenul/UBayFS/issues/23#issue-1482556305
But it seems that everything runs smoothly.
