+++ man/train.Rd
@@ -4,10 +4,12 @@
 \alias{train}
 \title{UBayFS feature selection}
 \usage{
-train(x)
+train(x, verbose = FALSE)
 }
 \arguments{
 \item{x}{a `UBaymodel` created by \link{build.UBaymodel}}
+
+\item{verbose}{if TRUE: GA optimization output is printed to the console}
 }
 \value{
 a `UBaymodel` with an additional list element output containing the optimized solution.
