### 2022-12-08 annajenul/UBayFS#23 @annajenul
Thank you very much for the comment! We included a "verbose" parameter in the train function and used message() to print the "Running Genetic Algorithm" as suggested.
