### 2022-12-08 annajenul/UBayFS#24 @dhvalden
Interactive Shiny app missing dependencies
Some the options listed in the Shiny app have missing dependencies not listed in the dependencies. The following in the message I got:
```
Error in { : task 1 failed - "there is no package called ‘randomForest’"
```
