### 2022-12-08 #4848 @dhvalden
I can confirm that the Shiny App works under my set-up. I found some small issues with the app, but they do not affect the main functionality:
- [ ] https://github.com/annajenul/UBayFS/issues/24#issue-1484910999

I can also confirm that the test functionality works and all the described tests passed with OK status.
Finally, I would like to see mentions in the paper of just a couple of examples of packages or tools with the same functionality, just to have a comparison. If this final point is addressed I would declare myself satisfied and close the review @osorensen.
