+++ R/trainUBaymodel.R
@@ -21,14 +21,18 @@ train.UBaymodel = function(x, verbose=FALSE){
   if(!is(x, "UBaymodel")){
     stop("Wrong class of model")
   }
-  if(is.null(x$constraint.params$constraints)){
+  if(is.null(x$constraint.params)){
     stop("At least a max-size constraint must be defined")
   }
-  ms = x$constraint.params$constraints$b[which(apply(x$constraint.params$constraints$A == 1, 1, all))]
+
+  # detect max-size constraint
+  feat_const <- which(sapply(x$constraint.params, function(y){return(identical(y$block_matrix, diag(nrow = ncol(x$data))))}))
+  ms = x$constraint.params[[feat_const]]$b[which(apply(x$constraint.params[[feat_const]]$A == 1, 1, all))]
+# ms = x$constraint.params[which(apply(x$constraint.params$constraints$A == 1, 1, all))]
   if((!is.numeric(ms)) || (length(ms) == 0)){
     stop("No max-size constraint among constraints")
   }
-  else if (ms > ncol(x$constraint.params$constraints$A)){
+  else if (ms > ncol(x$data)){
     stop("No max-size constraint among constraints")
   }
 
@@ -39,8 +43,7 @@ train.UBaymodel = function(x, verbose=FALSE){
     message("Running Genetic Algorithm")
     tGA <- train_GA(theta,
                          x$lambda,
-                         x$constraint.params$constraints,
-                         x$constraint.params$block_constraints,
+                         x$constraint.params,
                          x$optim.params,
                          colnames(x$data),
                          verbose)
@@ -77,14 +80,11 @@ train.UBaymodel = function(x, verbose=FALSE){
   return(x)
 }
 
-neg_loss <- function(state, theta, lambda, constraints, block_constraints){
+neg_loss <- function(state, theta, lambda, constraints){
   return(logSumExp(c(
           theta[state == 1],
-          log(lambda) + admissibility(state = state, # log( lambda * admissibility * block_admissibility )
-                                      constraints = constraints,
-                                      log = TRUE)
-                      + block_admissibility(state = state,
-                                      constraints = block_constraints,
+          log(lambda) + admissibility(state = state, # log( lambda * admissibility)
+                                      constraint_list = constraints,
                                       log = TRUE)
           )))}
 getNegLoss <- function(state, model, log = TRUE){
@@ -92,8 +92,7 @@ getNegLoss <- function(state, model, log = TRUE){
     state = state,
     theta = posteriorExpectation(model),
     lambda = model$lambda,
-    constraints = model$constraint.params$constraints,
-    block_constraints = model$constraint.params$block_constraints)
+    constraints = model$constraint.params)
   if(!log){
     res <- exp(res)
   }
@@ -101,18 +100,18 @@ getNegLoss <- function(state, model, log = TRUE){
 }
 
 
-train_GA <- function(theta, lambda, constraints, block_constraints, optim_params, feat_names, verbose){
+train_GA <- function(theta, lambda, constraints, optim_params, feat_names, verbose){
 
 
-  target_fct <- function(state){return(neg_loss(state, theta, lambda, constraints, block_constraints))}
+  target_fct <- function(state){return(neg_loss(state, theta, lambda, constraints))}
 
   # Greedy algorithm to select starting vectors
   x_start = sampleInitial(post_scores = exp(theta),
                           constraints = constraints,
-                          block_constraints = block_constraints,
                           size = optim_params$popsize)
 
 
+  print("SAMPLING FINISHED")
   optim = ga(type = "binary",								# use GA for optimization
              fitness = target_fct,
              lower = 0,
@@ -139,42 +138,33 @@ train_GA <- function(theta, lambda, constraints, block_constraints, optim_params
[…]
