### 2022-12-12 9dddd62
updated internal constraint structure
