+++ inst/shinyApp/UBayFSInteractive/server.R
@@ -12,589 +12,571 @@ library(caret)
 # Define server logic required to draw a histogram
 shinyServer(function(input, output, session) {
 
-    options(shiny.maxRequestSize=20*1024^2) # maximum upload size
+  options(shiny.maxRequestSize=20*1024^2) # maximum upload size
 
-    # help functions
-    addConstraint <- function(params, block_constraint = FALSE){
-      if(block_constraint){
-        newA <- params$A
-        newb <- params$b
-        newrho <- params$rho
+  # help functions
+  addConstraint <- function(params, block_constraint = FALSE){
+    if(block_constraint){
+      newA <- params$A
+      newb <- params$b
+      newrho <- params$rho
 
-        model(UBayFS::setBlockConstraints(model(),
-                                          list(A = newA, b = newb, rho = as.numeric(newrho), block_matrix = block_matrix()),
-                                          append = TRUE))
+      model(UBayFS::setConstraints(model(),
+                                   list(A = newA, b = newb, rho = as.numeric(newrho), block_matrix = block_matrix()),
+                                   append = TRUE))
 
-        proxy = dataTableProxy('blocks')
-        selectRows(proxy, c())
-      }
-      else{
-        newA <- params$A
-        newb <- params$b
-        newrho <- params$rho
-        colnames(newA) <- names_feats()
-
-        model(UBayFS::setConstraints(model(),
-                                     list(A = newA, b = newb, rho = as.numeric(newrho)),
-                                     append = TRUE))
-        proxy = dataTableProxy('features')
-        selectRows(proxy, c())
-      }
-
-
+      proxy = dataTableProxy('blocks')
+      selectRows(proxy, c())
     }
+    else{
+      newA <- params$A
+      newb <- params$b
+      newrho <- params$rho
+      colnames(newA) <- names_feats()
 
-    addWeights <- function(weights){
-      block_list = lapply(unique(blocks()), function(x){return(which(blocks() == x))})
-      model(setWeights(model(), weights, block_list = block_list))
-    }
-
-    # error messages
-    upload_error <- function(x){
-      show_alert(
-        title = "Error",
-        text = "An error occured during file upload - please check file and try again.",
-        type = 'error',
-        btn_labels = "Ok"
-      )
-    }
-
-    demo_error <- function(x){
-      show_alert(
-        title = "Error",
-        text = "An error occured during loading of demo file - please check internet connection and try again.",
-        type = 'error',
-        btn_labels = "Ok"
-      )
-    }
-
-    build_error <- function(x){
-      show_alert(
-        title = "Error",
-        text = "An error occured during application of UBayFS methods - please check data and parameters and try again.",
-        type = 'error',
-        btn_labels = "Ok"
-      )
-    }
-
-    # feature selection functions
-
-    decision_tree <- function(X, y, n, name = "tree"){
-      rf_data = as.data.frame(cbind(y, X))
-      colnames(rf_data) <- make.names(colnames(rf_data))
-      tree = rpart::rpart(y~., data = rf_data)
-      return(list(ranks= which(colnames(X) %in% names(tree$variable.importance)[1:n]),
-                  name = name))
-    }
-
-    rec_fe <- function(X,y,n, name="rfe"){
-      if(is.factor(y)){
-        control <- rfeControl(functions=rfFuncs, method = "cv", number = 2)
-      }
-      else{
-        control <- rfeControl(functions=lmFuncs, method = "cv", number = 2)
-      }
[…]
