### 2022-12-12 c1c8427
fixed issue with number of violated constraints
