+++ R/buildUBaymodel.R
@@ -66,12 +66,9 @@
 #'                             num_elements = length(bcw$blocks),
 #'                             rho = 10,
 #'                             block_list = bcw$blocks)
-#' model <- build.UBaymodel(
-#'                      data = bcw$data,
-#'                      target = bcw$labels,
-#'                      constraints = list(c, c_block),
-#'                      weights = w
-#' )
+#'
+#' model <- setConstraints(model, c_block)
+#'
 #' @import Rdimtools
 #' @import mRMRe
 #' @import shiny
