### 2022-12-13 a6b531c
rebuilt UBayconstraint as S3 class
