+++ man/is.UBayconstraint.Rd
@@ -1,13 +1,13 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/buildConstraints.R
-\name{checkConstraints}
-\alias{checkConstraints}
+\name{is.UBayconstraint}
+\alias{is.UBayconstraint}
 \title{Checks whether a list object implements proper UBayFS user constraints}
 \usage{
-checkConstraints(x)
+is.UBayconstraint(x)
 }
 \arguments{
-\item{x}{a `UBaymodel` object}
+\item{x}{a `UBayconstraint` object}
 }
 \value{
 boolean value
