+++ man/buildCustomConstraints.Rd
@@ -1,0 +1,23 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/buildConstraints.R
+\name{buildCustomConstraints}
+\alias{buildCustomConstraints}
+\title{Build a customized constraint for UBayFS}
+\usage{
+buildCustomConstraints(A, b, rho, block_matrix)
+}
+\arguments{
+\item{A}{matrix containing the left side of the linear inequality system}
+
+\item{b}{vector containing the right side of the linear inequality system}
+
+\item{rho}{vector containing the relaxation parameters for each constraint}
+
+\item{block_matrix}{a matrix indicating the membership of features in feature blocks}
+}
+\value{
+A `UBayconstraint` object
+}
+\description{
+Builds a constraint using a left side `A`, a right side `b`, a relaxation parameter `rho`, and a block matrix `block_matrix`.
+}
