### 2022-12-13 5f6309c
constraints as S3 class
