+++ NAMESPACE
@@ -8,9 +8,11 @@ export(admissibility)
 export(block_admissibility)
 export(build.UBaymodel)
 export(buildConstraints)
+export(buildCustomConstraints)
 export(buildDecorrConstraints)
 export(evaluateFS)
 export(evaluateMultiple)
+export(is.UBayconstraint)
 export(is.UBaymodel)
 export(printResults)
 export(runInteractive)
