### 2022-12-14 88e92cb
UBayconstraint class
