+++ man/build.UBayconstraint.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/buildConstraints.R
-\name{buildCustomConstraints}
-\alias{buildCustomConstraints}
+\name{build.UBayconstraint}
+\alias{build.UBayconstraint}
 \title{Build a customized constraint for UBayFS}
 \usage{
-buildCustomConstraints(A, b, rho, block_matrix)
+build.UBayconstraint(A, b, rho, block_matrix)
 }
 \arguments{
 \item{A}{matrix containing the left side of the linear inequality system}
