+++ NAMESPACE
@@ -6,9 +6,9 @@ S3method(summary,UBaymodel)
 S3method(train,UBaymodel)
 export(admissibility)
 export(block_admissibility)
+export(build.UBayconstraint)
 export(build.UBaymodel)
 export(buildConstraints)
-export(buildCustomConstraints)
 export(buildDecorrConstraints)
 export(evaluateFS)
 export(evaluateMultiple)
