+++ vignettes/paper.md
@@ -56,9 +56,9 @@ $$
 The optimization is implemented via a genetic algorithm along with a greedy algorithm for initialization, suggested by @jenul2022ubayfs to find a proper start vector for the optimization.
 
 # Package Summary
-The function `build.UBaymodel()` initializes an S3 class object UBaymodel and computes the ensemble of elementary feature selectors. In the current version, among others, linear feature selectors such as Fisher score and Laplacian score [@you2022rdimtools], mRMR [@de2013mrmre], or the non-linear HSIC Lasso are supported as options [@gselection]. In addition, the number of elementary models $M$ is specified. The user can directly set prior weights inside the `build` function. Constraints are either provided as a matrix $A$ and a right side $b$, or built using the `buildConstraints()` function, which supports max-size, must-link, and cannot-link constraints on both feature and block level. UBayFS requires at least one constraint limiting the total number of features to be selected ("max-size"). The level of constraint-relaxation is steered with an input parameter $\rho$. In addition, the weights for single features or feature blocks are set with `setWeights()`.
+The function `build.UBaymodel()` initializes an S3 class object UBaymodel and computes the ensemble of elementary feature selectors. In the current version, linear feature selectors such as Fisher score, Laplacian score [@you2022rdimtools], and mRMR [@de2013mrmre] are supported as integrated options. Any arbitrary feature selector can be defined manually and used as input. In addition, the number of elementary models $M$ is specified. The user can directly set prior weights inside the `build` function. Constraints are either provided as a matrix $A$ and a right side $b$, or built using the `buildConstraints()` function, which supports max-size, must-link, and cannot-link constraints on both feature and block level. UBayFS requires at least one constraint limiting the total number of features to be selected ("max-size"). The level of constraint-relaxation is steered with an input parameter $\rho$. In addition, the weights for single features or feature blocks are set with `setWeights()`.
 
-The functions `admissibility()` and `block_admissibility()` allow the user to evaluate the penalty term for a given feature set under a set of constraints. After initializing the model and computing the ensembles, the `train.UBaymodel()` function optimizes the feature set via a genetic algorithm [@scrucca2013ga] with greedy initialization. According to empirical evaluations, the greedy initialization decreases the runtime and leads to faster convergence towards an optimal feature set. Finally, the package implements the generic functions `print.UBaymodel()`, `plot.UBaymodel()`, and `summary.UBaymodel()` as well as an evaluation function `evaluteFS()` to report and visualize results. Two vignettes guide the user through the package and demonstrate how the method can be deployed in common application scenarios, including how user knowledge is specified and how feature- and block-wise constraints are set.
+The function `admissibility()`allows the user to evaluate the penalty term for a given feature set under a set of constraints. After initializing the model and computing the ensembles, the `train.UBaymodel()` function optimizes the feature set via a genetic algorithm [@scrucca2013ga] with greedy initialization. According to empirical evaluations, the greedy initialization decreases the runtime and leads to faster convergence towards an optimal feature set. Finally, the package implements the generic functions `print.UBaymodel()`, `plot.UBaymodel()`, and `summary.UBaymodel()` as well as an evaluation function `evaluteFS()` to report and visualize results. Two vignettes guide the user through the package and demonstrate how the method can be deployed in common application scenarios, including how user knowledge is specified and how feature- and block-wise constraints are set.
 
 # Interactive Shiny Dashboard
[…]
