+++ vignettes/BFS_UBayFS.Rmd
@@ -147,7 +147,7 @@ model <- setConstraints(model, constraints = block_constraints)
 summary(model)
 ```
 
-We observe, that the ``summary()`` prints an additional line for block constraints. All further steps are similar to the UBayFS variant without block constraints. In this example, three features from the third block are selected. Hence, all constraints are fulfilled.
+We observe, that the ``summary()`` prints additional lines for block constraints. All further steps are similar to the UBayFS variant without block constraints. In this example, three features from the third block are selected. Hence, all constraints are fulfilled.
 ```{r, include = TRUE}
 model = UBayFS::train(model)
 summary(model)
