+++ R/buildUBaymodel.R
@@ -78,7 +78,7 @@ build.UBaymodel = function(data,
                            target,
                            M = 100,
                            tt_split = 0.75,
-                           nr_features = 10,
+                           nr_features = "auto",
                            method = "mRMR",
                            prior_model = "dirichlet",
                            weights = 1,
