+++ man/print.UBayconstraint.Rd
@@ -1,0 +1,24 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/printUBayconstraint.R
+\name{print.UBayconstraint}
+\alias{print.UBayconstraint}
+\alias{summary.UBayconstraint}
+\title{Prints the `UBayconstraint` object}
+\usage{
+\method{print}{UBayconstraint}(x, ...)
+
+\method{summary}{UBayconstraint}(object, ...)
+}
+\arguments{
+\item{x}{a `UBayconstraint` object}
+
+\item{object}{a `UBayconstraint` object}
+}
+\description{
+Prints the `UBayconstraint` object
+}
+\section{Functions}{
+\itemize{
+\item \code{summary(UBayconstraint)}: Prints a summary of the `UBayconstraint` object
+
+}}
