+++ NAMESPACE
@@ -1,7 +1,9 @@
 # Generated by roxygen2: do not edit by hand
 
 S3method(plot,UBaymodel)
+S3method(print,UBayconstraint)
 S3method(print,UBaymodel)
+S3method(summary,UBayconstraint)
 S3method(summary,UBaymodel)
 S3method(train,UBaymodel)
 export(admissibility)
