### 2022-12-14 b48195b
integrated S3 class UBayconstraint
