### 2023-01-19 #4848 @EugeneHao
My issue has been resolved. The example and shiny app now work well on my machine.
