### 2023-01-19 annajenul/UBayFS#26 @osorensen
Reference formatting
This is an editorial comment to your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4848). This reference does not show as it should. The appropriate format would be "Jenul et al. (2022).

Here is the source:

https://github.com/annajenul/UBayFS/blob/208e2776e428e390baa0aeb83481112ef34b1781/vignettes/paper.bib#L12-L21

Could you try changing it to the following, and then running `@editorialbot generate pdf` in the review issue?

```
@article{jenul2022ubayfs,
  doi = {10.1007/s10994-022-06221-9},
  url = {https://doi.org/10.1007/s10994-022-06221-9},
  year = {2022},
  month = aug,
  publisher = {Springer Science and Business Media {LLC}},
  volume = {111},
  number = {10},
  pages = {3897--3923},
  author = {Anna Jenul and Stefan Schrunner and J\"{u}rgen Pilz and Oliver Tomic},
  title = {A user-guided Bayesian framework for ensemble feature selection in life science applications ({UBayFS})},
  journal = {Machine Learning}
}
```

One easy way of generating this bib entry is by entering 10.1007/s10994-022-06221-9 at https://www.doi2bib.org/.
