### 2023-01-19 annajenul/UBayFS#27 @osorensen
Reference formatting
This is another editorial comment on you [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4848).

The reference to the GSelection package does not appear properly. You can fix this by removing the extra set of braces around the `author` key:

https://github.com/annajenul/UBayFS/blob/208e2776e428e390baa0aeb83481112ef34b1781/vignettes/paper.bib#L57-L63

Could you try something like this?

```
@Manual{gselection,
    title = {{GSelection: Genomic Selection}},
    author = {Sayanti Guha Majumdar and Anil Rai and Dwijesh Chandra Mishra},
    year = {2019},
    note = {R package version 0.1.0},
    url = {https://CRAN.R-project.org/package=GSelection},
  }
```
