+++ vignettes/paper.bib
@@ -1,76 +1,88 @@
 @article{jenul2021rent,
-  title={{RENT—Repeated Elastic Net Technique for Feature Selection}},
-  author={Jenul, Anna and Schrunner, Stefan and Liland, Kristian Hovde and Indahl, Ulf Geir and Futsaether, Cecilia Marie and Tomic, Oliver},
-  journal={IEEE Access},
-  volume={9},
-  pages={152333--152346},
-  year={2021},
-  publisher={IEEE},
-  doi={10.1109/ACCESS.2021.3126429}
+  doi = {10.1109/access.2021.3126429},
+  url = {https://doi.org/10.1109/access.2021.3126429},
+  year = {2021},
+  publisher = {Institute of Electrical and Electronics Engineers ({IEEE})},
+  volume = {9},
+  pages = {152333--152346},
+  author = {Anna Jenul and Stefan Schrunner and Kristian Hovde Liland and Ulf Geir Indahl and Cecilia Marie Futsaether and Oliver Tomic},
+  title = {{RENT}{\textemdash}Repeated Elastic Net Technique for Feature Selection},
+  journal = {{IEEE} Access}
 }
 
 @article{jenul2022ubayfs,
-  title={{A User-Guided Bayesian Framework for Ensemble Feature Selection in Life Science Applications (UBayFS)}},
-  author={{Jenul, Anna and Schrunner, Stefan and Pilz, J{\"u}rgen and Tomic, Oliver}},
-  journal={{Machine Learning}},
-  pages={1--27},
-  year={2022},
-  publisher={{Springer}},
-  doi={10.1007/s10994-022-06221-9}
-
+  doi = {10.1007/s10994-022-06221-9},
+  url = {https://doi.org/10.1007/s10994-022-06221-9},
+  year = {2022},
+  month = aug,
+  publisher = {Springer Science and Business Media {LLC}},
+  volume = {111},
+  number = {10},
+  pages = {3897--3923},
+  author = {Anna Jenul and Stefan Schrunner and J\"{u}rgen Pilz and Oliver Tomic},
+  title = {A user-guided Bayesian framework for ensemble feature selection in life science applications ({UBayFS})},
+  journal = {Machine Learning}
 }
 
 @article{scrucca2013ga,
-  title={{GA: A Package for Genetic Algorithms in R}},
-  author={{Scrucca, Luca}},
-  journal={{Journal of Statistical Software}},
-  volume={53},
-  pages={1--37},
-  year={2013},
-  doi={10.18637/jss.v053.i04}
+  doi = {10.18637/jss.v053.i04},
+  url = {https://doi.org/10.18637/jss.v053.i04},
+  year = {2013},
+  publisher = {Foundation for Open Access Statistic},
+  volume = {53},
+  number = {4},
+  author = {Luca Scrucca},
+  title = {$\less$b$\greater${GA}$\less$/b$\greater$: A Package for Genetic Algorithms in$\less$i$\greater$R$\less$/i$\greater$},
+  journal = {Journal of Statistical Software}
 }
 
 @article{you2022rdimtools,
-  title={{Rdimtools: An R package for dimension reduction and intrinsic dimension estimation}},
-  author={{You, Kisung and Shung, Dennis}},
-  journal={{Software Impacts}},
-  pages={100414},
-  year={2022},
-  publisher={{Elsevier}},
-  doi={10.1016/j.simpa.2022.100414}
+  doi = {10.1016/j.simpa.2022.100414},
+  url = {https://doi.org/10.1016/j.simpa.2022.100414},
+  year = {2022},
+  month = dec,
+  publisher = {Elsevier {BV}},
+  volume = {14},
+  pages = {100414},
[…]
