+++ vignettes/paper.bib
@@ -32,7 +32,7 @@
   volume = {53},
   number = {4},
   author = {Luca Scrucca},
-  title = {$\less$b$\greater${GA}$\less$/b$\greater$: A Package for Genetic Algorithms in$\less$i$\greater$R$\less$/i$\greater$},
+  title = {{GA}: A Package for Genetic Algorithms in {R}},
   journal = {Journal of Statistical Software}
 }
 
@@ -45,7 +45,7 @@
   volume = {14},
   pages = {100414},
   author = {Kisung You and Dennis Shung},
-  title = {Rdimtools: An R package for dimension reduction and intrinsic dimension estimation},
+  title = {{Rdimtools}: An {R} package for dimension reduction and intrinsic dimension estimation},
   journal = {Software Impacts}
 }
 
@@ -60,7 +60,7 @@
   number = {18},
   pages = {2365--2368},
   author = {Nicolas De Jay and Simon Papillon-Cavanagh and Catharina Olsen and Nehme El-Hachem and Gianluca Bontempi and Benjamin Haibe-Kains},
-  title = {{mRMRe}: an R package for parallelized {mRMR} ensemble feature selection},
+  title = {{mRMRe}: an {R} package for parallelized {mRMR} ensemble feature selection},
   journal = {Bioinformatics}
 }
 
@@ -69,7 +69,7 @@
   title = {GSelection: Genomic Selection},
   author = {Sayanti Guha Majumdar and Anil Rai and Dwijesh Chandra Mishra},
   year = {2019},
-  note = {R package version 0.1.0},
+  note = {{R} package version 0.1.0},
   url = {https://CRAN.R-project.org/package=GSelection},
 }
 
@@ -81,7 +81,7 @@
   volume = {36},
   number = {11},
   author = {Miron B. Kursa and Witold R. Rudnicki},
-  title = {Feature Selection with the$\less$b$\greater$Boruta$\less$/b$\greater$Package},
+  title = {Feature Selection with the {Boruta} Package},
   journal = {Journal of Statistical Software}
 }
 
@@ -89,7 +89,7 @@
     title = {caret: Classification and Regression Training},
     author = {Max Kuhn},
     year = {2022},
-    note = {R package version 6.0-93},
+    note = {{R} package version 6.0-93},
     url = {https://CRAN.R-project.org/package=caret},
   }
 
