### 2023-01-19 annajenul/UBayFS#28 @osorensen
Mathematical notation
This is an editorial comment to your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4848).

In this equation, there appears to be a small typo. Shouldn't the dot "." be placed after the equation, and not after \kappa(t) (highlighted=?

Secondly, I'm not familiar with the notation used in the rightmost part of this equation. Does
mean "maximize the thing to the right with respect to delta, subject to delta in [0, 1]^N"?

If so, could you please add a sentence clarifying this?
