### 2023-01-19 annajenul/UBayFS#29 @osorensen
More reference formatting
This is an editorial comment to your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4848).

"Bayesian" should appear with a an uppercase B. Could you please fix this by adding a set of curly braces around the word here?

https://github.com/annajenul/UBayFS/blob/38a88f7aae07f2f85af6ee9150590f1871714a5e/vignettes/paper.bib#L23

That is, simply change the entry to `title = {A user-guided {Bayesian} framework for ensemble feature selection in life science applications ({UBayFS})}, `
