### 2023-01-19 annajenul/UBayFS#30 @osorensen
Citing R itself
This is an editorial comment to your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4848).

Could you please add a reference to R itself the first time it is mentioned, on line 13?

Issuing the `citation()` command inside R should give what you need:

```
> citation()

To cite R in publications use:

  R Core Team (2022). R: A language and environment for statistical computing. R Foundation for Statistical
  Computing, Vienna, Austria. URL https://www.R-project.org/.

A BibTeX entry for LaTeX users is

  @Manual{,
    title = {R: A Language and Environment for Statistical Computing},
    author = {{R Core Team}},
    organization = {R Foundation for Statistical Computing},
    address = {Vienna, Austria},
    year = {2022},
    url = {https://www.R-project.org/},
  }

We have invested a lot of time and effort in creating R, please cite it when using it for data analysis. See
also ‘citation("pkgname")’ for citing R packages.
```
