### 2023-01-19 annajenul/UBayFS#28 @annajenul
Thank you, good point. We rewrote the formulas and solved the issue with the "."
