+++ vignettes/paper.bib
@@ -20,7 +20,7 @@
   number = {10},
   pages = {3897--3923},
   author = {Anna Jenul and Stefan Schrunner and J\"{u}rgen Pilz and Oliver Tomic},
-  title = {A user-guided Bayesian framework for ensemble feature selection in life science applications ({UBayFS})},
+  title = {A user-guided {Bayesian} framework for ensemble feature selection in life science applications ({UBayFS})},
   journal = {Machine Learning}
 }
 
