+++ vignettes/paper.bib
@@ -50,6 +50,15 @@
 }
 
 
+@Manual{R,
+  title = {R: A Language and Environment for Statistical Computing},
+  author = {{R Core Team}},
+  organization = {R Foundation for Statistical Computing},
+  address = {Vienna, Austria},
+  year = {2022},
+  url = {https://www.R-project.org/},
+}
+
 @article{de2013mrmre,
   doi = {10.1093/bioinformatics/btt383},
   url = {https://doi.org/10.1093/bioinformatics/btt383},
