+++ vignettes/paper.md
@@ -33,7 +33,7 @@ The presented $\mathtt{R}$ package UBayFS provides an implementation along with 
 - mRMR
 
 $\mathtt{R}$ offers multiple packages implementing feature selection methodology. To name a few, $\mathtt{caret}$ [@caret] is an essential machine learning repository, containing models with built-in feature selection such as tree based methods (for instance $\mathtt{rpart2}$), regularized approaches like $\mathtt{lasso}$, and non-integrated feature selectors such as recursive feature elimination $\mathtt{rfe}$. Other examples are the $\mathtt{Boruta}$ [@boruta] package implementing the $\mathtt{Boruta}$ feature selector or the $\mathtt{GSelection}$ [@gselection] package containing $\mathtt{hsic\;lasso}$ feature selection. All feature selectors available in $\mathtt{R}$ can be used as underlying ensemble feature selectors in UBayFS. 
-Prior weights can be specified for single features or whole blocks as weight vectors. Linear side constraints are implemented via a matrix $A$ and a right side $b$ or with a customized function for specific constraint types. Hence, the sophisticated statistical model is summarized in a user-friendly and easy-to-use package.
+Prior weights can be specified for single features or whole blocks as weight vectors. Linear side constraints are implemented via a matrix $\boldsymbol{A}$ and a right side $\boldsymbol{b}$ or with a customized function for specific constraint types. Hence, the sophisticated statistical model is summarized in a user-friendly and easy-to-use package.
 
 
 ![At first, UBayFS elaborates information directly from data via ensemble feature selection. This information is merged with prior expert knowledge (a-priori feature weights) in a Bayesian model framework. Additionally, the user can include further side constraints such as a maximum number of features or cannot-link constraints between features. The final step comprises the optimization with respect to the model's utility function, including the side constraints.\label{fig:UBayFS}](UBayFS_concept.png)
@@ -44,7 +44,7 @@ As described in @jenul2022ubayfs, UBayFS is a Bayesian ensemble feature selectio
 
 $$p(\boldsymbol{\theta}|\boldsymbol{y})\propto p(\boldsymbol{y}|\boldsymbol{\theta})\cdot p(\boldsymbol{\theta}).$$
 
-In the framework of UBayFS,  $p(\boldsymbol{y}|\boldsymbol{\theta})$ represents the data-driven component (implemented via a multinomial likelihood), while $p(\boldsymbol{\theta})$ describes the user knowledge part modeled with a Dirichlet distribution. Due to the conjugate prior property of the Dirichlet distribution, the posterior parameter update has a tractable form and can be computed analytically.  Side constraints are represented by a system of linear inequalities $\boldsymbol{A}\cdot \boldsymbol{\delta}-\boldsymbol{b}\leq \boldsymbol{0}$, where $\boldsymbol{A}\in\mathbb{R}^{K\times N}$ and $\boldsymbol{b}\in\mathbb{R}^K$. $K$ is defined as the total number of constraints. 
[…]
