### 2023-01-19 9594a7a
correct typos in Shiny dashboard
