### 2023-01-20 #4848 @annajenul
- [ ] Make a tagged release of your software, and list the version tag of the archived version here: [v1.0.0](https://github.com/annajenul/UBayFS/releases/tag/v1.0.0)
- [ ] Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository): [https://zenodo.org/record/7554373#.Y8qHi6dBxFE](https://zenodo.org/record/7554373#.Y8qHi6dBxFE)
- [ ] Please list the DOI of the archived version here: [10.5281/zenodo.7554373](https://doi.org/10.5281/zenodo.7554373)
